mafia_game_langgraph/requirements.txt
a2a-sdk
dotenv
uvicorn
click
httpx
asyncio
nest_asyncio
pydantic
google-generativeai
langchain
langgraph